Supervise OneDHCPD.Server directly
This fixes an issue where `vintage_net` would fail to start a
`vintage_net_direct` interface when restoring its configuration due to
the OneDHCPD app not being started. Treating OneDHCPD as a library like
this fixes that issue.

This also has another benefit: Having the OneDHCPD.Server GenServer in
the VintageNet interface's supervision tree makes it easier to thinking
about what happens on failures on both OneDHCPD and VintageNetDirect.